refactor(ds): move the Field wiring into its own module (#443)
The controls this issue adds — Checkbox, RadioGroup, Switch — each need the
wiring a Field hands its control: the id, aria-describedby, aria-invalid,
aria-required and disabled. Reimplementing that per control is how the four
screens that predated Field ended up associating the label but not the error.

Exporting useControlWiring from Field.tsx does not work: a module exporting both
a component and a hook breaks Fast Refresh, and react-refresh/only-export-components
is at error level here. The rule is right, so the shared wiring gets its own
file, as the rule's own message suggests.

FieldStatus moves with it and is re-exported from Field.tsx, so
`import { type FieldStatus } from './Field'` keeps working — index.ts and the
call sites already spell it that way, and Résolution 1 of #443 freezes the
public API of the eight existing primitives. fieldContext is deliberately NOT
re-exported from index.ts: it is how this module builds controls, not something
a screen reaches for.

No behaviour change. The 34 tests in primitives.test.tsx pass unmodified.
